backend/requirements.prod.txt
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
django-filter==23.2
gunicorn==21.2.0
Markdown==3.4.4
Pillow==10.0.0
psycopg2-binary==2.9.7
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
sentry-sdk==1.30.0
PyJWT==2.8.0
